requirements.txt
python-dotenv==1.0.1
requests==2.32.3
google-api-python-client==2.141.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
